refactor: removed unused usings
