+++ TagsCloudResult/TagsCloudVisualization.WFApp/Actions/ImageSettingsAction.cs
@@ -1,5 +1,4 @@
 ﻿using TagsCloudVisualization.Common;
-using TagsCloudVisualization.Common.ResultOf;
 using TagsCloudVisualization.WFApp.Common;
 using TagsCloudVisualization.WFApp.Infrastructure;
