+++ TagsCloudResult/TagsCloudVisualization.WFApp/Actions/SourceSettingsAction.cs
@@ -1,5 +1,4 @@
 ﻿using TagsCloudVisualization.TextReaders;
-using TagsCloudVisualization.WFApp.Common;
 using TagsCloudVisualization.WFApp.Infrastructure;
 
 namespace TagsCloudVisualization.WFApp.Actions;
