+++ TagsCloudResult/TagsCloudVisualization/CloudLayouters/BaseCloudLayouter.cs
@@ -1,5 +1,4 @@
 ﻿using System.Drawing;
-using TagsCloudVisualization.Common;
 using TagsCloudVisualization.Common.ResultOf;
 using TagsCloudVisualization.PointsProviders;
