+++ TagsCloudResult/TagsCloudVisualization/CloudLayouters/ITagsCloudLayouter.cs
@@ -1,5 +1,4 @@
 ﻿using System.Drawing;
-using TagsCloudVisualization.Common;
 using TagsCloudVisualization.Common.ResultOf;
 
 namespace TagsCloudVisualization.CloudLayouters;
