+++ TagsCloudResult/TagsCloudVisualization/TextReaders/TxtTextReader.cs
@@ -1,5 +1,4 @@
-﻿using TagsCloudVisualization.Common;
-using TagsCloudVisualization.Common.ResultOf;
+﻿using TagsCloudVisualization.Common.ResultOf;
 
 namespace TagsCloudVisualization.TextReaders;
